refactor(device): break cycle + drop deprecated fullscreen aliases from device.* (#1467)
Two tweaks in response to PR feedback:

1. **Circular import broken** (Copilot review on fbf04aa08).
   `lang/deprecated.js` was importing `hasFullscreenSupport` /
   `isFullscreen` from `device.ts` while `device.ts` re-exported
   `requestFullscreen` / `exitFullscreen` from `lang/deprecated.js`.
   Factored both probes into a new leaf-level `system/fullscreen.ts`
   that owns the four-variant `DocumentLegacy` type and the
   detection logic; `device.ts` and `lang/deprecated.js` both import
   from there, eliminating the cycle.

2. **Dropped the `me.device.requestFullscreen` / `me.device.exit
   Fullscreen` aliases entirely.** `index.ts` already does
   `export * from "./lang/deprecated.js"`, so the two function names
   surface at the top level as `me.requestFullscreen()` /
   `me.exitFullscreen()` — a deprecated shim with runtime warning
   that points at `Application#requestFullscreen` /
   `Application#exitFullscreen`. Re-exporting them under
   `me.device.*` defeated the whole point of "they live in
   deprecated.js" — it leaked the deprecated names back into the
   namespace they were moved out of. Users still calling
   `me.device.requestFullscreen()` directly need to either migrate
   to `app.requestFullscreen()` (canonical) or drop the `.device.`
   segment.

`device.ts` keeps `hasFullscreenSupport` and `isFullscreen` exposed
under `me.device.*` via the new re-export from `system/fullscreen.ts`
— those are non-deprecated probes and remain valid public API.
